update(assets): create separate folder
