+++ src/components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import { LayoutProps } from "./types";
-import logo from "./logo.svg";
+import Logo from "../../assets/logo.svg";
 import "./styles.css";
 import { useStateContext } from "../context/StateContext";
 
@@ -10,7 +10,7 @@ export const Layout = ({ children }: LayoutProps) => {
     <div className="page">
       <div className="sidebar">
         <a href="/">
-          <img className="logo" src={logo} alt="Levo" />
+          <img className="logo" src={Logo} alt="Levo" />
         </a>
         {organization?.name && <p className="orgName">{organization.name}</p>}
         {organization?.name && <p className="pageName">Test Reports</p>}
